TurnPredict/requirements.txt
numpy
matplotlib.pyplot
json
csv
functools
polyline
random
pandas
tensorflow
tensorflow.contrib
sklearn 
googlemaps
urllib